chore: add missing type dependencies and fix ts casting issues

+++ frontend/src/utils/storage.ts
@@ -1,20 +1,23 @@
 import { userGetResponse } from "./interfaces"
 
+const isBrowser = typeof window !== 'undefined'
+
 const storage = {
     clear(): void {
-        window.sessionStorage.clear()
+        if(isBrowser) window.sessionStorage.clear()
     },
     jwt: {
         save(token: string): void {
-            window.sessionStorage.setItem(`jwt`, token)
+            if(isBrowser) window.sessionStorage.setItem(`jwt`, token)
         },
         load(bearer: boolean = true): string {
-            const token: string = window.sessionStorage.getItem(`jwt`) || ''
+            const token: string = isBrowser ? window.sessionStorage.getItem(`jwt`) || '' : ''
             if(bearer)
                 return `Bearer ${token}`
             return token
         },
         exists(): boolean {
+            if(!isBrowser) return false
             const token: string | null = window.sessionStorage.getItem(`jwt`)
             if(token === "null" || token === "undefined")
                 return false
@@ -23,9 +26,10 @@ const storage = {
     },
     user: {
         save(user: userGetResponse): void {
-            sessionStorage.setItem(`user`, JSON.stringify(user))
+            if(isBrowser) sessionStorage.setItem(`user`, JSON.stringify(user))
         },
         load(): userGetResponse | object {
+            if(!isBrowser) return {}
             try {
                 return JSON.parse(sessionStorage.getItem(`user`) as string)
             } catch(error: unknown) {
